requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
numpy==1.26.0
pandas==2.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
pyyaml==6.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
croniter==2.0.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.26.0
openpyxl==3.1.2
prometheus-client==0.19.0
structlog==24.1.0
